Re-organized REST server directory as model-view-controller
